Implemented first run at Drag & Drop
Edits made to the Drag and Drop function providing its base functionality.
Changed parseWorldX and parseWorldY to getWorldX and getWorldY, removing the need for an input double x and y. Instead, the getters for mouse coords are handled within the functions. They still return the same worldX and worldY mouse coordinate double values.

Should have some errors that will need to be hammered out in integration.